Refactor main into its own file
